django-docker/requirements.txt
asgiref==3.8.1
Django==5.0.10
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
psycopg==3.2.3
psycopg-binary==3.2.3
sqlparse==0.5.3
dataclasses-json==0.6.7
